Move some modules and merge some old work.
Try to keep the top level namespace more user-oriented.
Avoid having submodules depend on their parents.
