+++ src/scalems/exceptions.py
