+++ src/scalems/local/__main__.py
@@ -52,11 +52,11 @@ try:
             for name, ref in globals_namespace.items():
                 if isinstance(ref, scalems.core.ScriptEntryPoint):
                     if main is not None:
-                        raise scalems.core.exceptions.DispatchError('Multiple apps in the same script is not (yet?) supported.')
+                        raise scalems.exceptions.DispatchError('Multiple apps in the same script is not (yet?) supported.')
                     main = ref
                     ref.name = name
             if main is None:
-                raise scalems.core.exceptions.DispatchError('No scalems.app callables found in script.')
+                raise scalems.exceptions.DispatchError('No scalems.app callables found in script.')
             cmd = scalems.run(main, context=context)
         except SystemExit as e:
             exitcode = e.code
