+++ src/scalems/local_immediate/__init__.py
@@ -10,15 +10,9 @@ Example:
 
 """
 
-
-import asyncio
-import concurrent.futures
-import warnings
-from typing import Any, Callable
-
 import scalems.context
 from scalems.context import ItemView
-from scalems.core.exceptions import DuplicateKeyError, MissingImplementationError
+from scalems.exceptions import DuplicateKeyError, MissingImplementationError
 
 from . import operations
