+++ src/scalems/utility.py
@@ -1,36 +1,31 @@
-"""Core SCALE-MS implementation.
+"""Decorators and helper functions."""
 
-Represent the SCALE-MS object model in Python.
+__all__ = [
+    'app',
+    'command',
+    'function_wrapper',
+    'poll',
+    'run',
+    'wait',
+    'ScriptEntryPoint'
+]
 
-Specify and implement the core data model for Python user interfaces,
-middleware interfaces, and execution run time support.
-"""
-
-__all__ = ['app', 'run', 'wait', 'ScriptEntryPoint']
 
 import abc
-import contextlib
+import contextvars
 import functools
 import logging
 import typing
 import warnings
+from typing import Protocol
 
-import scalems.core.exceptions as exceptions
-from scalems.context import next_monotonic_integer
-from scalems.context import scope
+from scalems import exceptions
 
-from ..context import get_context
-from ..context import WorkflowManager
+from scalems.context import get_context, scope, WorkflowManager
+
 
 logger = logging.getLogger(__name__)
 logger.debug('Importing {}'.format(__name__))
-
-
-
-# TODO: Helpers and optimizations for fused operations, "partial" operations.
-# TODO: Distinguishing features of "dynamic" operations that can create new instances during execution.
-# TODO: Define Subgraph in terms of fused / partial operations.
-# TODO: Define "while_loop" in terms of dynamics operations.
 
 
 class ScriptEntryPoint(abc.ABC):
@@ -70,7 +65,80 @@ def app(func: typing.Callable) -> typing.Callable:
     return decorated
 
 
+
+# def command(*, input_type, result_type):
+#     """Get a decorator for ScaleMS Command definitions.
+#
+#     A ScaleMS command minimally consists of an input specification, and output
+#     specification, and a callable.
+#     """
+#     def decorator(cls):
+#         ...
+#     return decorator
+
+
+class Callable(Protocol):
+    """This protocol describes the required function signature for a SCALE-MS command."""
+    def __call__(self):
+        ...
+
+
+class Command(Protocol):
+    """Protocol describing a SCALE-MS Command."""
+
+
+def command(obj: Callable) -> Command:
+    """Decorate a callable to create a SCALE-MS Command."""
+    ...
+
+
+def function_wrapper(output: dict = None):
+    # Suppress warnings in the example code.
+    # noinspection PyUnresolvedReferences
+    """Generate a decorator for wrapped functions with signature manipulation.
+
+    New function accepts the same arguments, with additional arguments required by
+    the API.
+
+    The new function returns an object with an ``output`` attribute containing the named outputs.
+
+    Example:
+
+        >>> @function_wrapper(output={'spam': str, 'foo': str})
+        ... def myfunc(parameter: str = None, output=None):
+        ...    output.spam = parameter
+        ...    output.foo = parameter + ' ' + parameter
+        ...
+        >>> operation1 = myfunc(parameter='spam spam')
+        >>> assert operation1.spam.result() == 'spam spam'
+        >>> assert operation1.foo.result() == 'spam spam spam spam'
+
+    Arguments:
+        output (dict): output names and types
+
+    If ``output`` is provided to the wrapper, a data structure will be passed to
+    the wrapped functions with the named attributes so that the function can easily
+    publish multiple named results. Otherwise, the ``output`` of the generated operation
+    will just capture the return value of the wrapped function.
+    """
+    raise exceptions.MissingImplementationError()
+
+
+def poll():
+    """Inspect the execution status of an operation.
+
+    Inspects the execution graph state in the current context at the time of
+    execution.
+
+    Used in a work graph, this adds a non-deterministic aspect, but adds truly
+    asynchronous adaptability.
+    """
+    raise exceptions.MissingImplementationError()
+
+
 ResultType = typing.TypeVar('ResultType')
+
+
 class WorkflowObject(typing.Generic[ResultType]): ...
 
 
@@ -304,3 +372,24 @@ def run(work, context=None, **kwargs):
 
     # TODO: Consider generalized coroutines to be dispatched through
     #     custom event loops or executors.
+
+
+def next_monotonic_integer() -> int:
+    """Utility for generating a monotonic sequence of integers across an interpreter process.
+
+    Not thread-safe. However, threads may
+
+    * avoid race conditions by copying the contextvars context for non-root threads
+    * reproduce the sequence of the main thread by calling this function an equal
+      number of times.
+
+    Returns:
+        Next integer.
+
+    """
+    value = _monotonic_integer.get()
+    _monotonic_integer.set(value + 1)
+    return value
+
+
+_monotonic_integer = contextvars.ContextVar('_monotonic_integer', default=0)
